requirements.txt
streamlit
pandas
pyproj
google-generativeai